Instead of specifying numbers, you can now gamble by using percentages (of your total currency), all, half and mathematical equations - for example 5*5 or all-1 (one less than allin)

+++ NadekoBot.Core/Modules/Gambling/Common/AnimalRacing/AnimalRacingUser.cs
@@ -4,13 +4,13 @@ namespace NadekoBot.Modules.Gambling.Common.AnimalRacing
 {
     public class AnimalRacingUser
     {
-        public int Bet { get; }
+        public long Bet { get; }
         public string Username { get; }
         public ulong UserId { get; }
         public RaceAnimal Animal { get; set; }
         public int Progress { get; set; }
 
-        public AnimalRacingUser(string username, ulong userId, int bet)
+        public AnimalRacingUser(string username, ulong userId, long bet)
         {
             this.Bet = bet;
             this.Username = username;
